moved to amendable dir
fix docs
