+++ decidim-core/app/controllers/concerns/decidim/amendments_controller.rb
@@ -4,18 +4,18 @@ module Decidim
   class AmendmentsController < Decidim::ApplicationController
     include FormFactory
     before_action :authenticate_user!
-    helper_method :amendable
+    helper_method :amendable, :emendation
 
     def new
-      @form = form(Decidim::CreateAmendForm).from_model(amendable)
+      @form = form(Decidim::Amendable::CreateForm).from_model(amendable)
       @form.amendable_gid = params[:amendable_gid]
     end
 
     def create
-      @form = form(Decidim::CreateAmendForm).from_params(params)
+      @form = form(Decidim::Amendable::CreateForm).from_params(params)
       enforce_permission_to :create, :amend
 
-      Decidim::CreateAmend.call(@form) do
+      Decidim::Amendable::Create.call(@form) do
         on(:ok) do
           redirect_to Decidim::ResourceLocatorPresenter.new(@amendable).path
         end
@@ -28,37 +28,13 @@ module Decidim
     end
 
     def reject
-      return # to do!
-      @form = form(Decidim::RejectAmendForm).from_params(params)
-      enforce_permission_to :reject, :amend, amend: @form.amendable
+    end
 
-      Decidim::RejectAmend.call(@form) do
-        on(:ok) do
-          flash[:notice] = t("rejected.success", scope: "decidim.amendments")
-        end
-
-        on(:invalid) do
-          render json: { error: I18n.t("amendments.reject.error", scope: "decidim") }, status: 422
-        end
-      end
+    def review
+      @form = form(Decidim::Amendable::ReviewForm).from_model(emendation)
     end
 
     def accept
-      # # to do!
-      # @form = form(Decidim::AcceptAmendForm).from_params(params)
-      # enforce_permission_to :accept, :amend, amend: @form.amendable
-      #
-      # Decidim::AcceptAmend.call(@form, current_user) do
-      #   on(:ok) do
-      #     # flash[:notice] = t("accepted.success", scope: "decidim.amendments")
-      #   end
-      #
-      #   on(:invalid) do
-      #     # flash[:notice] = t("accepted.error", scope: "decidim.amendments")
-      #   end
-      #
-      #   redirect_to Decidim::ResourceLocatorPresenter.new(@emendation).path
-      # end
     end
 
     def amendable_gid
@@ -67,6 +43,14 @@ module Decidim
 
     def amendable
       @amendable ||= GlobalID::Locator.locate_signed amendable_gid
+    end
+
+    def emendation_gid
+      params[:emendation_gid]
+    end
+
+    def emendation
+      @emendation ||= GlobalID::Locator.locate_signed emendation_gid
     end
   end
 end
