+++ decidim-core/app/views/decidim/amendments/_form.html.erb
@@ -1,4 +1,4 @@
-<%= decidim_form_for(@form, url: amend_path, method: :post) do |form| %>
+<%= decidim_form_for(@form, url: amends_path, method: :post) do |form| %>
   <%= form.hidden_field :amendable_gid %>
   <%= form.hidden_field :amendable_type %>
   <%= form.hidden_field :emendation_type %>
