+++ decidim-core/config/routes.rb
@@ -104,7 +104,11 @@ Decidim::Core::Engine.routes.draw do
 
   resource :follow, only: [:create, :destroy]
   resource :report, only: [:create]
-  resource :amend, only: [:new, :create, :reject, :accept], controller: :amendments
+  resources :amends, only: [:new], controller: :amendments do
+    collection do
+      post :create
+    end
+  end
 
   namespace :gamification do
     resources :badges, only: [:index]
